Attempting to add missing id functions
